Analyze jumps from l.j and l.jr
For l.jr it is required to know contents of the register. Usually, if it
is not used as return, this contents is set by a pair of instructions:
l.movhi and l.ori. In this two cases contents is saved, so it is
available when l.jr needs it.

For the purpose of following state of registers, two variables are
introduced - one is array of register values, the other is enabling the
register reading, when register is set to known value